refactor: move appendSnapshot to own util function
