+++ src/CuteChat.tsx
@@ -12,6 +12,7 @@ import React, {
 import { Alert, NativeScrollEvent } from 'react-native';
 import type { IMessage } from 'react-native-gifted-chat';
 import { GiftedChat, GiftedChatProps } from 'react-native-gifted-chat';
+import { appendSnapshot } from './utils/appendSnapshot';
 
 interface CustomCuteChatProps {
   chatId: string;
@@ -30,7 +31,7 @@ interface User {
 type CuteChatProps = Omit<GiftedChatProps, 'messages' | 'user' | 'onSend'> &
   CustomCuteChatProps;
 
-type SnapshotChange = {
+export type SnapshotChange = {
   type: FirebaseFirestore.DocumentChangeType;
   message: IMessage;
 };
@@ -389,48 +390,5 @@ export const prepareSnapshot = async (
       type: change.type,
       message: await docToMessage(change.doc, chatId),
     }))
-  );
-};
-
-export const appendSnapshot = (
-  currentMessages: IMessage[],
-  snapshotChanges: SnapshotChange[]
-): IMessage[] => {
-  console.log('Appending snapshot...');
-  const newMessages = [...currentMessages];
-
-  for (const change of snapshotChanges) {
-    switch (change.type) {
-      case 'removed':
-        console.log('Message remove id:', change.message._id);
-        const deleteIndex = newMessages.findIndex(
-          (m) => change.message._id === m._id
-        );
-        if (deleteIndex === -1) {
-          console.log('Message does not exist in currentMessage');
-          break;
-        }
-        newMessages.splice(deleteIndex, 1);
-        console.log('Message removed');
-        break;
-      case 'added':
-      case 'modified':
-        console.log('Message modified or added id:', change.message._id);
-        const modifiedIndex = newMessages.findIndex(
-          (m) => change.message._id === m._id
-        );
-        if (modifiedIndex === -1) {
-          console.log('Message added');
-          newMessages.push(change.message);
-        } else {
-          console.log('Message updated');
-          newMessages[modifiedIndex] = change.message;
-        }
-        break;
-    }
-  }
-
-  return newMessages.sort(
-    (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
   );
 };
